+++ picodata-plugin/src/error_code.rs
@@ -82,8 +82,9 @@ tarantool::define_enum_with_introspection! {
         /// Picodata machinery is not yet initialized on the instance.
         Uninitialized = 10024,
 
-        /// Attempt to expel an instance which is not currently Offline.
-        ExpelOnlineInstance = 10025,
+        /// Instance is not allowed to be expelled in the given situation for
+        /// some reason. The most often solution is to use `picodata expel --force`.
+        ExpelNotAllowed = 10025,
 
         /// Not an actual error code, just designates the start of the range.
         UserDefinedErrorCodesStart = 20000,