+++ test/int/test_ddl.py
@@ -563,7 +563,8 @@ def test_ddl_create_table_partial_failure(cluster: Cluster):
     assert entry[4][0] == "ddl_prepare"
 
     # Expel the last conflicting instance to fix the conflict.
-    i1.call("pico.expel", i5.name)
+    # TODO(https://git.picodata.io/core/picodata/-/issues/1100) picodata nuke
+    i1.call("pico.expel", i5.name, dict(force=True))
     applied_index = i1.call("box.space._raft_state:get", "applied")[1]
 
     # After that expel we expect a ddl commit